spaceinv01: bring all aliens in a class, if a peng hits an alien, the alien is dead........
If the alien is at the boundary from the playground the go all a step down.
If a peng hits the ship, the ship is destroy.
Add a randomgenerator
Want to make that the aliens can shoot.
If the y axis of the point bl from a alien aabbox lesser -62, it is game over.