spaceinv: per level the aliens shooted more and the pengs will be faster
If all aliens are dead, a new level started.
The aliens shooted randomly.
If a peng hits a bunker, than the peng is destroied.
There are max. 3 pengs from the ship on the screen.
The pengs are slower.
There are lives for the ship. If the live number is zero, it`s game over.
There is a game over text and if it is game over the text will be draw.